requirements.txt
pyautogui
opencv-python
numpy
Pillow 